requirements.txt
streamlit==1.17.0
requests==2.28.2
beautifulsoup4==4.11.1
google-generativeai==0.3.0
python-dotenv==0.21.0
langchain==0.0.108
faiss-cpu==1.7.2
sentence-transformers==2.2.0
